+++ frontend/sozisel-app/src/components/SessionDetails/CreateSession.scss
@@ -1,6 +1,6 @@
 @import "../../styles/styles";
 
-.SessionCreation {
+.CreateSession {
     height: calc(100% - #{$navbar-height});
     background-color: $bg-01;
     margin: 0;
