+++ frontend/sozisel-app/src/components/SessionDetails/EditSession/EditSession.scss
@@ -1,6 +1,6 @@
 @import "../../../styles/styles";
 
-.SessionEditing {
+.EditSession {
     height: calc(100% - #{$navbar-height});
     background-color: $bg-01;
     margin: 0;
