+++ frontend/sozisel-app/src/components/SessionDetails/SessionDetails/SessionDetails.scss
@@ -1,8 +1,8 @@
 @import "../../../styles/styles";
 
-.SessionDetailsContainer {
+.SessionDetails {
     height: 85%;
-    .sessionDetails {
+    .sessionDetailsForm {
         padding: $spacing-04;
         border-bottom: 1px solid $bg-01;
